refactor: move newClient into mongodeb folder
