+++ app/utils/models/getDocument.ts
@@ -1,4 +1,4 @@
-import { client } from "./newClient"
+import { client } from "../mongodb/newClient"
 
 async function getDocument() {
   try {
